vector-store: move filter value JSON parsing into `cql_types`
Pure move of `try_from_json` into `cql_types::from_json`.
Behavior stays unchanged. Both JSON directions now live together for easier
cross-checking.
