+++ crates/vector-store/src/httproutes.rs
@@ -34,7 +34,6 @@ use crate::vector;
 use crate::vs_index;
 use crate::vs_index::VsIndexSearch;
 use crate::vs_index::VsIndexSearchExt;
-use anyhow::anyhow;
 use anyhow::bail;
 use axum::Router;
 use axum::extract;
@@ -51,7 +50,6 @@ use axum::response::Response;
 use axum::routing::get;
 use axum::routing::put;
 use axum_server_dual_protocol::Protocol;
-use bigdecimal::BigDecimal;
 use httpapi::DataType;
 use httpapi::FulltextIndexOptions;
 use httpapi::IndexInfo;
@@ -59,29 +57,16 @@ use httpapi::IndexOptions;
 use httpapi::SimilarityFunction;
 use httpapi::VectorIndexOptions;
 use itertools::Itertools;
-use num_bigint::BigInt;
 use prometheus::Encoder;
 use prometheus::ProtobufEncoder;
 use prometheus::TextEncoder;
-use regex::Regex;
 use scylla::cluster::metadata::NativeType;
-use scylla::value::CqlDecimal;
-use scylla::value::CqlTimeuuid;
 use scylla::value::CqlValue;
-use scylla::value::CqlVarint;
 use serde_json::Value;
 use std::collections::HashMap;
-use std::num::NonZero;
 use std::num::NonZeroUsize;
 use std::sync::Arc;
-use std::sync::LazyLock;
 use std::sync::RwLock;
-use time::Date;
-use time::OffsetDateTime;
-use time::Time;
-use time::format_description::well_known::Iso8601;
-use time::format_description::well_known::iso8601::Config;
-use time::format_description::well_known::iso8601::TimePrecision;
 use tokio::sync::mpsc::Sender;
 use tower_http::trace::TraceLayer;
 use tracing::debug;
@@ -1229,7 +1214,7 @@ fn try_from_post_index_ann_filter(
                 "Column '{column}' in filter restriction is not part of the table or is not a supported native type",
             )
         };
-        try_from_json(value, native_type)
+        cql_types::from_json(value, native_type)
     };
     Ok(Filter {
         restrictions: json_filter
@@ -1418,158 +1403,6 @@ fn try_collect_primary_keys(
         .collect()
 }
 
-fn try_from_json(value: Value, cql_type: &NativeType) -> anyhow::Result<CqlValue> {
-    match value {
-        Value::String(value) => match cql_type {
-            NativeType::Ascii => Ok(CqlValue::Ascii(value)),
-            NativeType::Text => Ok(CqlValue::Text(value)),
-            NativeType::Uuid => {
-                let uuid = value
-                    .parse()
-                    .map_err(|err| anyhow!("Failed to parse UUID from string '{value}': {err}"))?;
-                Ok(CqlValue::Uuid(uuid))
-            }
-            NativeType::Timeuuid => {
-                let timeuuid: CqlTimeuuid = value.parse().map_err(|err| {
-                    anyhow!("Failed to parse TimeUUID from string '{value}': {err}")
-                })?;
-                Ok(CqlValue::Timeuuid(timeuuid))
-            }
-            NativeType::Date => {
-                let date = Date::parse(&value, &Iso8601::DATE)
-                    .map_err(|err| anyhow!("Failed to parse Date from string '{value}': {err}"))?;
-                Ok(CqlValue::Date(date.into()))
-            }
-            NativeType::Time => {
-                let time = Time::parse(value.strip_prefix("T").unwrap_or(&value), &Iso8601::TIME)
-                    .map_err(|err| {
-                    anyhow!("Failed to parse Time from string '{value}': {err}")
-                })?;
-                Ok(CqlValue::Time(time.into()))
-            }
-            NativeType::Timestamp => {
-                // CQL timestamps may use a space as the date-time separator
-                // (e.g. '2024-01-01 00:00:00.000Z'), but ISO 8601 requires 'T'.
-                // Only normalize when the space occurs at the expected date-time
-                // boundary after a YYYY-MM-DD prefix; otherwise, leave the input
-                // unchanged so that error reporting reflects the original value.
-                static CQL_TIMESTAMP_RE: LazyLock<Regex> =
-                    LazyLock::new(|| Regex::new(r"^(\d{4}-\d{2}-\d{2}) ").expect("valid regex"));
-                let normalized = CQL_TIMESTAMP_RE.replace(&value, "${1}T");
-                let datetime = OffsetDateTime::parse(&normalized, {
-                    const CONFIG: u128 = Config::DEFAULT
-                        .set_time_precision(TimePrecision::Second {
-                            decimal_digits: NonZero::new(3),
-                        })
-                        .encode();
-                    &Iso8601::<CONFIG>
-                })
-                .map_err(|err| anyhow!("Failed to parse Timestamp from string '{value}': {err}"))?;
-                Ok(CqlValue::Timestamp(datetime.into()))
-            }
-            NativeType::Blob => {
-                if !value.starts_with("0x") {
-                    bail!("Blob value must be a '0x'-prefixed hex string");
-                }
-                let bytes = const_hex::decode(&value)
-                    .map_err(|err| anyhow!("Invalid hex in blob value: {err}"))?;
-                Ok(CqlValue::Blob(bytes))
-            }
-            NativeType::Varint => {
-                let bi: BigInt = value.parse().map_err(|err| {
-                    anyhow!("Failed to parse Varint from string '{value}': {err}")
-                })?;
-                Ok(CqlValue::Varint(CqlVarint::from(bi)))
-            }
-            NativeType::Decimal => {
-                let bd: BigDecimal = value.parse().map_err(|err| {
-                    anyhow!("Failed to parse Decimal from string '{value}': {err}")
-                })?;
-                Ok(CqlValue::Decimal(CqlDecimal::try_from(bd).map_err(
-                    |err| anyhow!("Decimal value out of range: {err}"),
-                )?))
-            }
-            _ => bail!("Cannot convert string to CqlValue::{cql_type:?}, unsupported type"),
-        },
-
-        Value::Bool(value) => match cql_type {
-            NativeType::Boolean => Ok(CqlValue::Boolean(value)),
-            _ => bail!("Cannot convert bool to CqlValue::{cql_type:?}, unsupported type"),
-        },
-        Value::Number(value) => match cql_type {
-            NativeType::Double => {
-                Ok(CqlValue::Double(value.as_f64().ok_or_else(|| {
-                    anyhow!("Expected f64 for CqlValue::Double")
-                })?))
-            }
-            NativeType::Float => {
-                Ok(CqlValue::Float({
-                    // there is no TryFrom<f64> for f32, so we use explicit conversion
-                    let value = value
-                        .as_f64()
-                        .ok_or_else(|| anyhow!("Expected f32 (type f64) for CqlValue::Float"))?;
-                    if !value.is_finite() || value < f32::MIN as f64 || value > f32::MAX as f64 {
-                        bail!("Expected f32 for CqlValue::Float: value out of range");
-                    }
-                    let value = value as f32;
-                    if !value.is_finite() {
-                        bail!("Expected finite f32 for CqlValue::Float");
-                    }
-                    value
-                }))
-            }
-            NativeType::Int => Ok(CqlValue::Int(
-                value
-                    .as_i64()
-                    .ok_or_else(|| anyhow!("Expected i32 (type i64) for CqlValue::Int"))?
-                    .try_into()
-                    .map_err(|err| anyhow!("Expected i32 for CqlValue::Int: {err}"))?,
-            )),
-            NativeType::BigInt => {
-                Ok(CqlValue::BigInt(value.as_i64().ok_or_else(|| {
-                    anyhow!("Expected i64 for CqlValue::BigInt")
-                })?))
-            }
-            NativeType::SmallInt => Ok(CqlValue::SmallInt(
-                value
-                    .as_i64()
-                    .ok_or_else(|| anyhow!("Expected i16 (type i64) for CqlValue::SmallInt"))?
-                    .try_into()
-                    .map_err(|err| anyhow!("Expected i16 for CqlValue::SmallInt: {err}"))?,
-            )),
-            NativeType::TinyInt => Ok(CqlValue::TinyInt(
-                value
-                    .as_i64()
-                    .ok_or_else(|| anyhow!("Expected i8 (type i64) for CqlValue::TinyInt"))?
-                    .try_into()
-                    .map_err(|err| anyhow!("Expected i8 for CqlValue::TinyInt: {err}"))?,
-            )),
-            NativeType::Varint => {
-                // Varint is always an integer; reject fractional JSON numbers.
-                let s = value.to_string();
-                let bi: BigInt = s
-                    .parse()
-                    .map_err(|err| anyhow!("Failed to parse Varint from number '{s}': {err}"))?;
-                Ok(CqlValue::Varint(CqlVarint::from(bi)))
-            }
-            NativeType::Decimal => {
-                let s = value.to_string();
-                let bd: BigDecimal = s
-                    .parse()
-                    .map_err(|err| anyhow!("Failed to parse Decimal from number '{s}': {err}"))?;
-                Ok(CqlValue::Decimal(CqlDecimal::try_from(bd).map_err(
-                    |err| anyhow!("Decimal value out of range: {err}"),
-                )?))
-            }
-            _ => bail!("Cannot convert number to CqlValue::{cql_type:?}, unsupported type"),
-        },
-
-        _ => {
-            bail!("Cannot convert JSON value '{value}' to CqlValue::{cql_type:?}, unsupported type")
-        }
-    }
-}
-
 #[utoipa::path(
     get,
     path = "/api/v1/info",
@@ -1676,8 +1509,6 @@ mod tests {
 
     use super::*;
     use crate::Analyzer;
-    use serde_json::Number;
-    use uuid::Uuid;
 
     #[test]
     fn try_from_post_index_ann_filter_conversion_ok() {
@@ -1968,250 +1799,6 @@ mod tests {
                 &table_columns
             )
             .is_err()
-        );
-    }
-
-    #[test]
-    fn try_from_json_conversion() {
-        assert_eq!(
-            try_from_json(Value::String("ascii".to_string()), &NativeType::Ascii).unwrap(),
-            CqlValue::Ascii("ascii".to_string())
-        );
-        assert_eq!(
-            try_from_json(Value::String("text".to_string()), &NativeType::Text).unwrap(),
-            CqlValue::Text("text".to_string())
-        );
-
-        assert_eq!(
-            try_from_json(Value::Bool(true), &NativeType::Boolean).unwrap(),
-            CqlValue::Boolean(true)
-        );
-
-        assert_eq!(
-            try_from_json(
-                Value::Number(Number::from_f64(101.).unwrap()),
-                &NativeType::Double
-            )
-            .unwrap(),
-            CqlValue::Double(101.)
-        );
-        assert_eq!(
-            try_from_json(
-                Value::Number(Number::from_f64(201.).unwrap()),
-                &NativeType::Float
-            )
-            .unwrap(),
-            CqlValue::Float(201.)
-        );
-        assert!(
-            try_from_json(
-                Value::Number(Number::from_f64((f32::MAX as f64) * 10.).unwrap()),
-                &NativeType::Float
-            )
-            .is_err()
-        );
-        assert!(
-            try_from_json(
-                Value::Number(Number::from_f64((f32::MIN as f64) * 10.).unwrap()),
-                &NativeType::Float
-            )
-            .is_err()
-        );
-
-        assert_eq!(
-            try_from_json(Value::Number(10.into()), &NativeType::Int).unwrap(),
-            CqlValue::Int(10)
-        );
-        assert!(
-            try_from_json(
-                Value::Number((i32::MAX as i64 + 1).into()),
-                &NativeType::Int
-            )
-            .is_err()
-        );
-        assert_eq!(
-            try_from_json(Value::Number(20.into()), &NativeType::BigInt).unwrap(),
-            CqlValue::BigInt(20)
-        );
-        assert_eq!(
-            try_from_json(Value::Number(30.into()), &NativeType::SmallInt).unwrap(),
-            CqlValue::SmallInt(30)
-        );
-        assert!(
-            try_from_json(
-                Value::Number((i16::MAX as i64 + 1).into()),
-                &NativeType::SmallInt
-            )
-            .is_err()
-        );
-        assert_eq!(
-            try_from_json(Value::Number(40.into()), &NativeType::TinyInt).unwrap(),
-            CqlValue::TinyInt(40)
-        );
-        assert!(
-            try_from_json(
-                Value::Number((i8::MAX as i64 + 1).into()),
-                &NativeType::TinyInt
-            )
-            .is_err()
-        );
-
-        let uuid = Uuid::new_v4();
-        assert_eq!(
-            try_from_json(Value::String(uuid.into()), &NativeType::Uuid).unwrap(),
-            CqlValue::Uuid(uuid)
-        );
-        let uuid = Uuid::new_v4();
-        assert_eq!(
-            try_from_json(Value::String(uuid.into()), &NativeType::Timeuuid).unwrap(),
-            CqlValue::Timeuuid(uuid.into())
-        );
-
-        assert_eq!(
-            try_from_json(Value::String("2025-09-01".to_string()), &NativeType::Date).unwrap(),
-            CqlValue::Date(
-                Date::from_calendar_date(2025, time::Month::September, 1)
-                    .unwrap()
-                    .into()
-            )
-        );
-        assert_eq!(
-            try_from_json(
-                Value::String("12:10:10.000000000".to_string()),
-                &NativeType::Time
-            )
-            .unwrap(),
-            CqlValue::Time(Time::from_hms(12, 10, 10).unwrap().into())
-        );
-        assert_eq!(
-            try_from_json(
-                Value::String(
-                    // truncate microseconds
-                    OffsetDateTime::from_unix_timestamp(123456789)
-                        .unwrap()
-                        .format({
-                            const CONFIG: u128 = Config::DEFAULT
-                                .set_time_precision(TimePrecision::Second {
-                                    decimal_digits: NonZero::new(3),
-                                })
-                                .encode();
-                            &Iso8601::<CONFIG>
-                        })
-                        .unwrap()
-                ),
-                &NativeType::Timestamp
-            )
-            .unwrap(),
-            CqlValue::Timestamp(
-                OffsetDateTime::from_unix_timestamp(123456789)
-                    .unwrap()
-                    .into()
-            )
-        );
-
-        // CQL-style timestamp with space separator and Z offset
-        assert_eq!(
-            try_from_json(
-                Value::String("2024-01-01 00:00:00.000Z".to_string()),
-                &NativeType::Timestamp
-            )
-            .unwrap(),
-            CqlValue::Timestamp(
-                OffsetDateTime::from_unix_timestamp(1704067200)
-                    .unwrap()
-                    .into()
-            )
-        );
-
-        // CQL-style timestamp with space separator, Z offset, and non-zero time
-        assert_eq!(
-            try_from_json(
-                Value::String("1970-01-01 00:01:04.000Z".to_string()),
-                &NativeType::Timestamp
-            )
-            .unwrap(),
-            CqlValue::Timestamp(OffsetDateTime::from_unix_timestamp(64).unwrap().into())
-        );
-
-        assert_eq!(
-            try_from_json(Value::String("0xdeadbeef".to_string()), &NativeType::Blob).unwrap(),
-            CqlValue::Blob(vec![0xde, 0xad, 0xbe, 0xef])
-        );
-        assert_eq!(
-            try_from_json(Value::String("0x".to_string()), &NativeType::Blob).unwrap(),
-            CqlValue::Blob(vec![])
-        );
-        assert_eq!(
-            try_from_json(Value::String("0x00".to_string()), &NativeType::Blob).unwrap(),
-            CqlValue::Blob(vec![0x00])
-        );
-
-        // missing 0x prefix
-        assert!(try_from_json(Value::String("deadbeef".to_string()), &NativeType::Blob).is_err());
-        // invalid hex characters
-        assert!(try_from_json(Value::String("0xgg".to_string()), &NativeType::Blob).is_err());
-        // odd-length hex digits (after stripping prefix)
-        assert!(try_from_json(Value::String("0xabc".to_string()), &NativeType::Blob).is_err());
-
-        // Varint from string
-        assert_eq!(
-            try_from_json(
-                Value::String("-98765432109876543210987654321098765432109876543210".to_string()),
-                &NativeType::Varint
-            )
-            .unwrap(),
-            CqlValue::Varint(CqlVarint::from(
-                "-98765432109876543210987654321098765432109876543210"
-                    .parse::<BigInt>()
-                    .unwrap()
-            ))
-        );
-        assert!(
-            try_from_json(
-                Value::String("not_a_number".to_string()),
-                &NativeType::Varint
-            )
-            .is_err()
-        );
-        // Varint from JSON number
-        assert_eq!(
-            try_from_json(Value::Number((-9876543210i64).into()), &NativeType::Varint).unwrap(),
-            CqlValue::Varint(CqlVarint::from(BigInt::from(-9876543210i64)))
-        );
-
-        // Decimal from string
-        assert_eq!(
-            try_from_json(
-                Value::String("-98765432109876543210.123456789".to_string()),
-                &NativeType::Decimal
-            )
-            .unwrap(),
-            CqlValue::Decimal(
-                CqlDecimal::try_from(
-                    "-98765432109876543210.123456789"
-                        .parse::<BigDecimal>()
-                        .unwrap()
-                )
-                .unwrap()
-            )
-        );
-        assert!(
-            try_from_json(
-                Value::String("not_a_decimal".to_string()),
-                &NativeType::Decimal
-            )
-            .is_err()
-        );
-        // Decimal from JSON number
-        assert_eq!(
-            try_from_json(
-                Value::Number(Number::from_f64(-1.25).unwrap()),
-                &NativeType::Decimal
-            )
-            .unwrap(),
-            CqlValue::Decimal(
-                CqlDecimal::try_from("-1.25".parse::<BigDecimal>().unwrap()).unwrap()
-            )
         );
     }
